+++ packages/client-sdk-nodejs/src/internal/vector-index-control-client.ts
@@ -6,23 +6,18 @@ import {Status} from '@grpc/grpc-js/build/src/constants';
 import {cacheServiceErrorMapper} from '../errors/cache-service-error-mapper';
 import {ChannelCredentials, Interceptor} from '@grpc/grpc-js';
 import {
-  CreateCache,
-  DeleteCache,
   ListCaches,
   CreateSigningKey,
   ListSigningKeys,
   RevokeSigningKey,
-  CacheFlush,
   CredentialProvider,
   MomentoLogger,
-  CacheInfo,
+  VectorIndexConfiguration,
 } from '..';
 import {version} from '../../package.json';
 import {IdleGrpcClientWrapper} from './grpc/idle-grpc-client-wrapper';
 import {GrpcClientWrapper} from './grpc/grpc-client-wrapper';
-import {Configuration} from '../config/configuration';
 import {
-  validateCacheName,
   validateIndexName,
   validateNumDimensions,
   validateTtlMinutes,
@@ -30,21 +25,18 @@ import {
 import {normalizeSdkError} from '@gomomento/sdk-core/dist/src/errors';
 import {_SigningKey} from '@gomomento/sdk-core/dist/src/messages/responses/grpc-response-types';
 import {
-  CacheLimits,
-  TopicLimits,
-} from '@gomomento/sdk-core/dist/src/messages/cache-info';
-import {
   CreateVectorIndex,
   DeleteVectorIndex,
   ListVectorIndexes,
 } from '@gomomento/sdk-core';
+import {IVectorIndexControlClient} from '@gomomento/sdk-core/dist/src/internal/clients';
 
 export interface ControlClientProps {
-  configuration: Configuration;
+  configuration: VectorIndexConfiguration;
   credentialProvider: CredentialProvider;
 }
 
-export class ControlClient {
+export class VectorIndexControlClient implements IVectorIndexControlClient {
   private readonly clientWrapper: GrpcClientWrapper<grpcControl.ScsControlClient>;
   private readonly interceptors: Interceptor[];
   private static readonly REQUEST_TIMEOUT_MS: number = 60 * 1000;
@@ -61,7 +53,7 @@ export class ControlClient {
     ];
     this.interceptors = [
       new HeaderInterceptorProvider(headers).createHeadersInterceptor(),
-      ClientTimeoutInterceptor(ControlClient.REQUEST_TIMEOUT_MS),
+      ClientTimeoutInterceptor(VectorIndexControlClient.REQUEST_TIMEOUT_MS),
     ];
     this.logger.debug(
       `Creating control client using endpoint: '${props.credentialProvider.getControlEndpoint()}`
@@ -76,129 +68,6 @@ export class ControlClient {
       maxIdleMillis: props.configuration
         .getTransportStrategy()
         .getMaxIdleMillis(),
-    });
-  }
-
-  public async createCache(name: string): Promise<CreateCache.Response> {
-    try {
-      validateCacheName(name);
-    } catch (err) {
-      return new CreateCache.Error(normalizeSdkError(err as Error));
-    }
-    this.logger.info(`Creating cache: ${name}`);
-    const request = new grpcControl._CreateCacheRequest({
-      cache_name: name,
-    });
-    return await new Promise<CreateCache.Response>(resolve => {
-      this.clientWrapper.getClient().CreateCache(
-        request,
-        {interceptors: this.interceptors},
-        // eslint-disable-next-line @typescript-eslint/no-unused-vars
-        (err, resp) => {
-          if (err) {
-            if (err.code === Status.ALREADY_EXISTS) {
-              resolve(new CreateCache.AlreadyExists());
-            } else {
-              resolve(new CreateCache.Error(cacheServiceErrorMapper(err)));
-            }
-          } else {
-            resolve(new CreateCache.Success());
-          }
-        }
-      );
-    });
-  }
-
-  public async deleteCache(name: string): Promise<DeleteCache.Response> {
-    try {
-      validateCacheName(name);
-    } catch (err) {
-      return new DeleteCache.Error(normalizeSdkError(err as Error));
-    }
-    const request = new grpcControl._DeleteCacheRequest({
-      cache_name: name,
-    });
-    this.logger.info(`Deleting cache: ${name}`);
-    return await new Promise<DeleteCache.Response>(resolve => {
-      this.clientWrapper.getClient().DeleteCache(
-        request,
-        {interceptors: this.interceptors},
-        // eslint-disable-next-line @typescript-eslint/no-unused-vars
-        (err, resp) => {
-          if (err) {
-            resolve(new DeleteCache.Error(cacheServiceErrorMapper(err)));
-          } else {
-            resolve(new DeleteCache.Success());
-          }
-        }
-      );
-    });
-  }
-
-  public async flushCache(cacheName: string): Promise<CacheFlush.Response> {
-    try {
-      validateCacheName(cacheName);
-    } catch (err) {
-      return new CacheFlush.Error(normalizeSdkError(err as Error));
-    }
-    this.logger.trace(`Flushing cache: ${cacheName}`);
-    return await this.sendFlushCache(cacheName);
-  }
-
-  private async sendFlushCache(
-    cacheName: string
-  ): Promise<CacheFlush.Response> {
-    const request = new grpcControl._FlushCacheRequest({
-      cache_name: cacheName,
-    });
-    return await new Promise(resolve => {
-      this.clientWrapper.getClient().FlushCache(
-        request,
-        {
-          interceptors: this.interceptors,
-        },
-        (err, resp) => {
-          if (resp) {
-            resolve(new CacheFlush.Success());
-          } else {
-            resolve(new CacheFlush.Error(cacheServiceErrorMapper(err)));
-          }
-        }
-      );
-    });
-  }
-
-  public async listCaches(): Promise<ListCaches.Response> {
-    const request = new grpcControl._ListCachesRequest();
-    request.next_token = '';
-    this.logger.debug("Issuing 'listCaches' request");
-    return await new Promise<ListCaches.Response>(resolve => {
-      this.clientWrapper
-        .getClient()
-        .ListCaches(request, {interceptors: this.interceptors}, (err, resp) => {
-          if (err || !resp) {
-            resolve(new ListCaches.Error(cacheServiceErrorMapper(err)));
-          } else {
-            const caches = resp.cache.map(cache => {
-              const cacheName = cache.cache_name;
-              const topicLimits: TopicLimits = {
-                maxPublishMessageSizeKb:
-                  cache.topic_limits?.max_publish_message_size_kb || 0,
-                maxSubscriptionCount:
-                  cache.topic_limits?.max_subscription_count || 0,
-                maxPublishRate: cache.topic_limits?.max_publish_rate || 0,
-              };
-              const cacheLimits: CacheLimits = {
-                maxTtlSeconds: cache.cache_limits?.max_ttl_seconds || 0,
-                maxItemSizeKb: cache.cache_limits?.max_item_size_kb || 0,
-                maxThroughputKbps: cache.cache_limits?.max_throughput_kbps || 0,
-                maxTrafficRate: cache.cache_limits?.max_traffic_rate || 0,
-              };
-              return new CacheInfo(cacheName, topicLimits, cacheLimits);
-            });
-            resolve(new ListCaches.Success(caches));
-          }
-        });
     });
   }
