+++ packages/client-sdk-nodejs/src/preview-vector-index-client.ts
@@ -1,14 +1,12 @@
-import {ControlClient} from './internal/control-client';
 import {
   AbstractVectorIndexClient,
   IVectorIndexClient,
   IVectorIndexControlClient,
 } from '@gomomento/sdk-core/dist/src/internal/clients';
 import {VectorIndexClientProps} from './vector-index-client-props';
-import {VectorDataClient} from './internal/vector-data-client';
+import {VectorIndexDataClient} from './internal/vector-index-data-client';
+import {VectorIndexControlClient} from './internal/vector-index-control-client';
 import {IVectorIndexDataClient} from '@gomomento/sdk-core/dist/src/internal/clients/vector/IVectorIndexDataClient';
-import {CacheConfiguration} from './config/configuration';
-import {defaultRetryStrategy} from './config/configurations';
 
 /**
  * PREVIEW Vector Index Client
@@ -32,15 +30,8 @@ export class PreviewVectorIndexClient
 function createControlClient(
   props: VectorIndexClientProps
 ): IVectorIndexControlClient {
-  return new ControlClient({
-    configuration: new CacheConfiguration({
-      loggerFactory: props.configuration.getLoggerFactory(),
-      transportStrategy: props.configuration.getTransportStrategy(),
-      retryStrategy: defaultRetryStrategy(
-        props.configuration.getLoggerFactory()
-      ),
-      middlewares: [],
-    }),
+  return new VectorIndexControlClient({
+    configuration: props.configuration,
     credentialProvider: props.credentialProvider,
   });
 }
@@ -48,5 +39,5 @@ function createControlClient(
 function createDataClient(
   props: VectorIndexClientProps
 ): IVectorIndexDataClient {
-  return new VectorDataClient(props);
+  return new VectorIndexDataClient(props);
 }
